requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-heroku==0.3.1
django-import-export==1.2.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==2.6.2
pipenv==2018.11.26
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.3
Werkzeug==0.15.2
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
